iot_dashboard/requirements.txt
Flask
paho-mqtt
influxdb-client
telegraf